Special string
Keep count of the chars. If the count involves more than one char and both chars count is greater that one then its impossible for the string to be valid.
If the total count is the same as the count of any char in the string then they are all the same.
If the count in an odd lengted string minus the mid char is the same as the count of any char in the string then they are all the same except the one mid char so its valid.